# docs/openapi3.yaml
openapi: 3.0.0
...
info:
  contact:
    email: shpakovskiipa@gmail.com
    name: API Support
  description: This is a Eschool API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Eschool API
  version: "1.0"
paths:
...
  /schools:
    get:
      description: get all schools
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.SchoolDTO"
                type: array
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: GetAllSchools
      tags:
        - school
    post:
      description: create school
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                http_v1_dto.CreateSchoolDTO"
        description: created school info
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                  http_v1_dto.SchoolDTO"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: CreateSchool
      tags:
        - school
  "/schools/{courseID}/courses/{lessonID}":
    delete:
      description: delete course lesson
      parameters:
        - description: course id
          in: path
          name: courseID
          required: true
          schema:
            type: string
        - description: lesson id
          in: path
          name: lessonID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: message
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: DeleteCourseLesson
      tags:
        - course
...
  "/schools/{id}":
    get:
      description: get school by id
      parameters:
        - description: school id
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.SchoolDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: GetSchoolByID
      tags:
        - school
    put:
      description: update school
      parameters:
        - description: school id
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                http_v1_dto.UpdateSchoolDTO"
        description: updated school info
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                  http_v1_dto.SchoolDTO"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: UpdateSchool
      tags:
        - school
  "/schools/{id}/courses":
    get:
      description: get school courses
      parameters:
        - description: school id
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.CourseDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: GetSchoolCourses
      tags:
        - school
    post:
      description: create school course
      parameters:
        - description: school id
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                http_v1_dto.CreateCourseDTO"
        description: created course info
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.CourseDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: CreateSchoolCourse
      tags:
        - school
  "/schools/{id}/teachers":
    get:
      description: get school teachers
      parameters:
        - description: school id
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                    http_v1_dto.UserDTO"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      summary: GetSchoolTeachers
      tags:
        - school
  "/schools/{schoolID}/courses/{courseID}":
    delete:
      description: delete school course
      parameters:
        - description: school id
          in: path
          name: schoolID
          required: true
          schema:
            type: string
        - description: course id
          in: path
          name: courseID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: message
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: DeleteSchoolCourse
      tags:
        - school
    put:
      description: update school course
      parameters:
        - description: school id
          in: path
          name: schoolID
          required: true
          schema:
            type: string
        - description: course id
          in: path
          name: courseID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                http_v1_dto.UpdateCourseDTO"
        description: updated course info
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_paw1a_eschool_internal_adapter_delivery_\
                  http_v1_dto.CourseDTO"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: UpdateSchoolCourse
      tags:
        - school
  "/schools/{schoolID}/teachers/{teacherID}":
    put:
      description: add school teacher
      parameters:
        - description: school id
          in: path
          name: schoolID
          required: true
          schema:
            type: string
        - description: course id
          in: path
          name: teacherID
          required: true
          schema:
            type: string
      responses:
        "201":
          description: message
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorBadReque\
                  st"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorUnauthor\
                  ized"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorNotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_adapter_delivery_http_v1.RestErrorInternal\
                  Error"
      security:
        - ApiKeyAuth: []
      summary: AddSchoolTeacher
      tags:
        - school
...
servers:
  - url: //localhost:8080/api/v1
components:
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    internal_adapter_delivery_http_v1.RestErrorNotFound:
      properties:
        error:
          example: not found
          type: string
        status:
          example: 404
          type: integer
        timestamp:
          example: 2020-11-10T23:00:00+00:00
          type: string
      type: object
...